refactor(server): extract daemon start/stop helpers
Moves _confirmShutdown / _startDaemon / _refreshWhenDaemonStops out
of dashboard_screen.dart into a shared server_actions.dart so the
new Server screen can reuse them. Adds a restartDaemon() helper for
the Listen URL editor's restart banner.

Dashboard's existing call sites are unchanged thanks to one-line
forwarders.

Refs #397
