+++ flutter_app/lib/features/dashboard/dashboard_screen.dart
@@ -4,7 +4,6 @@ import 'package:go_router/go_router.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import '../../core/models/pr.dart';
 import '../../core/models/tracked_issue.dart';
-import '../../core/platform/platform_services_provider.dart';
 import '../../shared/widgets/severity_badge.dart';
 import '../../shared/widgets/state_badge.dart';
 import '../../shared/widgets/toast.dart';
@@ -21,6 +20,7 @@ import '../stats/stats_screen.dart';
 import 'activity_filter_bar.dart';
 import 'activity_filters.dart';
 import 'dashboard_providers.dart';
+import '../server/server_actions.dart' as server_actions;
 
 class DashboardScreen extends ConsumerWidget {
   const DashboardScreen({super.key});
@@ -113,115 +113,11 @@ class DashboardScreen extends ConsumerWidget {
   }
 }
 
-const _daemonStartHealthMaxAttempts = 80;
-const _daemonStartHealthInterval = Duration(milliseconds: 100);
+Future<void> _confirmShutdown(BuildContext context, WidgetRef ref) =>
+    server_actions.confirmShutdown(context, ref);
 
-Future<void> _confirmShutdown(BuildContext context, WidgetRef ref) async {
-  final confirmed = await showDialog<bool>(
-    context: context,
-    builder: (context) => AlertDialog(
-      title: const Text('Stop Server?'),
-      content: const Text('Active reviews may be interrupted.'),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.of(context).pop(false),
-          child: const Text('Cancel'),
-        ),
-        FilledButton(
-          onPressed: () => Navigator.of(context).pop(true),
-          child: const Text('Stop Server'),
-        ),
-      ],
-    ),
-  );
-  if (confirmed != true || !context.mounted) return;
-
-  try {
-    await ref.read(apiClientProvider).shutdownDaemon();
-    if (!context.mounted) return;
-    showToast(context, 'Shutdown requested');
-    ref.invalidate(sseStreamProvider);
-    ref.invalidate(daemonHealthProvider);
-    await _refreshWhenDaemonStops(context, ref);
-  } catch (e) {
-    if (context.mounted) showToast(context, 'Error: $e', isError: true);
-  }
-}
-
-Future<void> _startDaemon(BuildContext context, WidgetRef ref) async {
-  if (ref.read(daemonStartingProvider)) return;
-
-  final platform = ref.read(platformServicesProvider);
-  final binaryPath = platform.defaultDaemonBinaryPath();
-  if (binaryPath == null || binaryPath.isEmpty) {
-    showToast(context, 'Daemon binary not found', isError: true);
-    return;
-  }
-
-  ref.read(daemonStartingProvider.notifier).state = true;
-  try {
-    await platform.spawnDaemon(binaryPath);
-    final api = ref.read(apiClientProvider);
-    var healthy = false;
-    for (var i = 0; i < _daemonStartHealthMaxAttempts; i++) {
-      await Future<void>.delayed(_daemonStartHealthInterval);
-      healthy = await api.checkHealth();
-      if (healthy) break;
-    }
-    _invalidateDashboardData(ref);
-    if (!context.mounted) return;
-    if (healthy) {
-      showToast(context, 'Server started');
-    } else {
-      showToast(
-        context,
-        'Heimdallm could not start. Check the app installation.',
-        isError: true,
-      );
-    }
-  } catch (e) {
-    if (context.mounted) showToast(context, 'Error: $e', isError: true);
-  } finally {
-    ref.read(daemonStartingProvider.notifier).state = false;
-  }
-}
-
-Future<void> _refreshWhenDaemonStops(
-  BuildContext context,
-  WidgetRef ref,
-) async {
-  final api = ref.read(apiClientProvider);
-  const delays = [
-    Duration(milliseconds: 200),
-    Duration(milliseconds: 300),
-    Duration(milliseconds: 500),
-    Duration(milliseconds: 800),
-    Duration(milliseconds: 1200),
-    Duration(seconds: 2),
-  ];
-
-  for (final delay in delays) {
-    await Future<void>.delayed(delay);
-    if (!context.mounted) return;
-    final healthy = await api.checkHealth();
-    if (!context.mounted) return;
-    ref.invalidate(daemonHealthProvider);
-    if (!healthy) break;
-  }
-
-  if (!context.mounted) return;
-  _invalidateDashboardData(ref);
-}
-
-void _invalidateDashboardData(WidgetRef ref) {
-  ref.invalidate(sseStreamProvider);
-  ref.invalidate(daemonHealthProvider);
-  ref.invalidate(prsProvider);
-  ref.invalidate(issuesProvider);
-  ref.invalidate(statsProvider);
-  ref.invalidate(activityEntriesProvider);
-  ref.invalidate(activityOptionsProvider);
-}
+Future<void> _startDaemon(BuildContext context, WidgetRef ref) =>
+    server_actions.startDaemon(context, ref);
 
 // ── Reviews tab ──────────────────────────────────────────────────────────────
